requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
app==0.0.1
asttokens==3.0.0
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
chex==0.1.88
click==8.1.8
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
coverage==7.6.10
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.11
decorator==5.1.1
docker-pycreds==0.4.0
etils==1.5.2
executing==2.1.0
fastai==2.7.18
fastapi==0.115.7
fastcore==1.7.28
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.16.1
flax==0.10.2
fonttools==4.55.5
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
hydra-core==1.3.2
idna==3.10
imageio==2.36.1
imbalanced-learn==0.13.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
-e git+https://github.com/MrCogito/mlops_dtu_2025_project7.git@2b8e4bc0b2960e325404afed2dba7cba3aa179d3#egg=mlops_2025_floods_prediction
mpmath==1.3.0
msgpack==1.1.0
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.0
pyparsing==3.2.1
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
pyts==0.13.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.0
sentry-sdk==2.20.0
setproctitle==1.3.4
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.2
sympy==1.13.1
tensorstore==0.1.71
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2024.12.12
toolz==1.0.0
torch==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.2.0
tsai==0.3.9
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
zipp==3.21.0
